Module lattice_qcd_rs::error

defines different error types.

Structs

ErrorWithOnwnedValue
A struct that combine an error with a owned value

Enums

ImplementationError
Errors in the implementation of the library. This is unwanted to return this type but somethimes this is better to return that instead of panicking.

LatticeInitializationError
Error while initialising a lattice

MultiIntegrationError
Error return while doing multiple steps.

Never
Type that can never be (safly) initialized. This is temporary, until never is accepted into stable rust.

StateInitializationError
Error while initialising a state

StateInitializationErrorThreaded
Error while initialising a state