Struct lattice_qcd_rs::field::EField

pub struct EField<const D: usize> { /* fields omitted */ }

Represent an electric field.

Implementations

impl<const D: usize> EField<D>

pub fn new(data: Vec<SVector<Su3Adjoint, D>>) -> Self

Create a new “Electrical” field.

pub const fn data(&self) -> &Vec<SVector<Su3Adjoint, D>>

Get the raw data.

pub fn len(&self) -> usize

Return the number of elements.

pub fn new_deterministe(
l: &LatticeCyclique<D>,
rng: &mut impl Rng,
d: &impl Distribution<Real>
) -> Self

Single threaded generation with a given random number generator. useful to reproduce a set of data.

Example

extern crate rand;
extern crate rand_distr;
use rand::{rngs::StdRng, SeedableRng};

let mut rng_1 = StdRng::seed_from_u64(0);
let mut rng_2 = StdRng::seed_from_u64(0);
// They have the same seed and should generate the same numbers
let distribution = rand::distributions::Uniform::from(-1_f64..1_f64);
let lattice = LatticeCyclique::<4>::new(1_f64, 4).unwrap();
assert_eq!(
    EField::new_deterministe(&lattice, &mut rng_1, &distribution),
    EField::new_deterministe(&lattice, &mut rng_2, &distribution)
);

pub fn new_random(l: &LatticeCyclique<D>, d: &impl Distribution<Real>) -> Self

Single thread generation by seeding a new rng number. To create a seedable and reproducible set use EField::new_deterministe..

pub fn new_cold(l: &LatticeCyclique<D>) -> Self

Create a new cold configuration for the electriccal field, i.e. all E ar set to 0.

pub fn get_e_vec(
&self,
point: &LatticePoint<D>,
l: &LatticeCyclique<D>
) -> Option<&SVector<Su3Adjoint, D>>

Get E(point) = [E_x(point), E_y(point), E_z(point)].

pub fn get_e_field(
&self,
point: &LatticePoint<D>,
dir: &Direction<D>,
l: &LatticeCyclique<D>
) -> Option<&Su3Adjoint>

Get E_{dir}(point). The sign of the direction does not change the output. i.e. E_{-dir}(point) = E_{dir}(point).

pub fn is_empty(&self) -> bool

Returns wether there is no data

pub fn get_gauss(
&self,
link_matrix: &LinkMatrix,
point: &LatticePoint<D>,
lattice: &LatticeCyclique<D>
) -> Option<CMatrix3>

Return the Gauss parameter G(x) = \sum_i E_i(x) - U_{-i}(x) E_i(x - i) U^\dagger_{-i}(x).

pub fn get_gauss_sum_div(
&self,
link_matrix: &LinkMatrix,
lattice: &LatticeCyclique<D>
) -> Option<Real>

Get the deviation from the Gauss law

pub fn project_to_gauss(
&self,
link_matrix: &LinkMatrix,
lattice: &LatticeCyclique<D>
) -> Option<Self>

project to that the gauss law is approximatively respected ( up to f64::EPSILON * 10 per point)