Struct lattice_qcd_rs::field::Su3Adjoint

pub struct Su3Adjoint { /* fields omitted */ }

Adjoint representation of SU(3), it is su(3) (i.e. the lie algebra). See su3::GENERATORS to view the order of generators. Note that the generators are normalize such that Tr[T^a T^b] = \delta^{ab} / 2

Implementations

impl Su3Adjoint

pub const fn new(data: Vector8<Real>) -> Self

create a new Su3Adjoint representation where M = M^a T^a, where T are generators given in su3::GENERATORS.

Example

extern crate nalgebra;
use lattice_qcd_rs::field::Su3Adjoint;

let su3 = Su3Adjoint::new(nalgebra::SVector::<f64, 8>::from_element(1_f64));

pub fn new_from_array(data: [Real; 8]) -> Self

create a new Su3Adjoint representation where M = M^a T^a, where T are generators given in su3::GENERATORS.

Example

let su3 = Su3Adjoint::new_from_array([0_f64; 8]);

pub const fn data(&self) -> &Vector8<Real>

get the data inside the Su3Adjoint.

pub fn to_matrix(&self) -> Matrix3<Complex<Real>>

return the su(3) (Lie algebra) matrix.

Example

let su3 = Su3Adjoint::new_from_array([1_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64]);
assert_eq!(su3.to_matrix(), *lattice_qcd_rs::su3::GENERATORS[0]);

pub fn to_su3(self) -> Matrix3<Complex<Real>>

Return the SU(3) matrix associated with this generator. Note that the function consume self.

Example

let su3 = Su3Adjoint::new_from_array([1_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64, 0_f64]);
assert_eq!(su3.to_su3().determinant(), nalgebra::Complex::from(1_f64));

pub fn exp(self) -> Matrix3<Complex<Real>>

return exp( T^a v^a) where v is self. Note that the function consume self.

pub fn random(rng: &mut impl Rng, d: &impl Distribution<Real>) -> Self

create a new random SU3 adjoint.

Example

extern crate rand;
extern crate rand_distr;

let mut rng = rand::thread_rng();
let distribution = rand::distributions::Uniform::from(-1_f64..1_f64);
let su3 = Su3Adjoint::random(&mut rng, &distribution);

pub fn trace_squared(&self) -> Real

Returns the trace squared Tr(X^2).

It is more accurate and faster than computing

(m.to_matrix() * m.to_matrix()).trace().real();

pub fn t(&self) -> Real

Return the t coeff t = - 1/2 * Tr(X^2). If you are looking for the trace square use Self::trace_squared instead.

It is used for su3::su3_exp_i.

Example

let su3 = Su3Adjoint::from([1_f64; 8]);
let m = su3.to_matrix();
assert_eq!(
    nalgebra::Complex::new(su3.t(), 0_f64),
    -nalgebra::Complex::from(0.5_f64) * (m * m).trace()
);

pub fn d(&self) -> Complex<Real>

Return the t coeff d = i * det(X). Used for su3::su3_exp_i

Example

let su3 = Su3Adjoint::from([1_f64; 8]);
let m = su3.to_matrix();
assert_eq!(
    su3.d(),
    nalgebra::Complex::new(0_f64, 1_f64) * m.determinant()
);

pub fn len(&self) -> usize

Return the number of data. This number is 8

assert_eq!(su3.len(), 8);

pub fn iter(&self) -> impl Iterator<Item = &Real>

Get an iterator over the ellements.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut Real>

Get an iterator over the mutable ref of ellements.

pub fn data_mut(&mut self) -> &mut Vector8<Real>

Get a mutlable reference over the data.