Module lattice_qcd_rs::integrator

Numerical integrators to carry out simulations.

See SymplecticIntegrator. The simulations are done on LatticeStateWithEField It also require a notion of SimulationStateSynchrone and SimulationStateLeapFrog.

Even thought it is effortless to implement both SimulationStateSynchrone and SimulationStateLeapFrog. I adivce against it and implement only SimulationStateSynchrone and use super::simulation::SimulationStateLeap for leap frog states as it gives a compile time check that you did not forget doing a demi steps.

This library gives two implementations of SymplecticIntegrator: SymplecticEuler and SymplecticEulerRayon. I would advice using SymplecticEulerRayon if you do not mind the little more momory it uses.

Example

let us create a basic random state and let us simulate.

extern crate rand;
extern crate rand_distr;
use lattice_qcd_rs::integrator::SymplecticEuler;
use lattice_qcd_rs::simulation::LatticeStateDefault;
use lattice_qcd_rs::simulation::LatticeStateWithEField;
use lattice_qcd_rs::simulation::LatticeStateWithEFieldSyncDefault;
use lattice_qcd_rs::simulation::SimulationStateSynchrone;

let mut rng = rand::thread_rng();
let distribution =
    rand::distributions::Uniform::from(-std::f64::consts::PI..std::f64::consts::PI);
let state1 = LatticeStateWithEFieldSyncDefault::new_random_e_state(
    LatticeStateDefault::<3>::new_deterministe(100_f64, 1_f64, 4, &mut rng).unwrap(),
    &mut rng,
);
let state2 = state1
    .simulate_sync(&SymplecticEuler::new(8), 0.0001_f64)
    .unwrap();
let state3 = state2
    .simulate_sync(&SymplecticEuler::new(8), 0.0001_f64)
    .unwrap();

Let us then compute and compare the Hamiltonian.

let h = state1.get_hamiltonian_total();
let h2 = state3.get_hamiltonian_total();
println!("The error on the Hamiltonian is {}", h - h2);

Re-exports

pub use symplectic_euler::SymplecticEuler;

pub use symplectic_euler_rayon::SymplecticEulerRayon;

Modules

symplectic_euler
Basic symplectic Euler integrator

symplectic_euler_rayon
Basic symplectic Euler integrator using Rayon.

Traits

SymplecticIntegrator
Define an symplectic numerical integrator