Trait lattice_qcd_rs::integrator::SymplecticIntegrator

pub trait SymplecticIntegrator<StateSync, StateLeap, const D: usize> where
    StateSync: SimulationStateSynchrone<D>,
    StateLeap: SimulationStateLeapFrog<D>, {
    type Error;
    fn integrate_sync_sync(
        &self, 
        l: &StateSync, 
        delta_t: Real
    ) -> Result<StateSync, Self::Error>;

    fn integrate_leap_leap(
        &self, 
        l: &StateLeap, 
        delta_t: Real
    ) -> Result<StateLeap, Self::Error>;

    fn integrate_sync_leap(
        &self, 
        l: &StateSync, 
        delta_t: Real
    ) -> Result<StateLeap, Self::Error>;

    fn integrate_leap_sync(
        &self, 
        l: &StateLeap, 
        delta_t: Real
    ) -> Result<StateSync, Self::Error>;

    fn integrate_symplectic(
        &self, 
        l: &StateSync, 
        delta_t: Real
    ) -> Result<StateSync, Self::Error> { ... }
}

Define an symplectic numerical integrator

The integrator evlove the state in time.

The integrator should be capable of switching between Sync state (q (or link matrices) at time T , p (or e_field) at time T ) and leap frog (a at time T, p at time T + 1/2)

Associated Types

type Error

Type of error returned by the Integrator.

Loading content...
Required methods

fn integrate_sync_sync(
&self,
l: &StateSync,
delta_t: Real
) -> Result<StateSync, Self::Error>

Integrate a sync state to a sync state.

Errors

Return an error if the integration encounter a problem

fn integrate_leap_leap(
&self,
l: &StateLeap,
delta_t: Real
) -> Result<StateLeap, Self::Error>

Integrate a leap state to a leap state using leap frog algorithm.

Errors

Return an error if the integration encounter a problem

fn integrate_sync_leap(
&self,
l: &StateSync,
delta_t: Real
) -> Result<StateLeap, Self::Error>

Integrate a sync state to a leap state by doing a half step for the conjugate momenta.

Errors

Return an error if the integration encounter a problem

fn integrate_leap_sync(
&self,
l: &StateLeap,
delta_t: Real
) -> Result<StateSync, Self::Error>

Integrate a leap state to a sync state by finishing doing a step for the position and finishing the half step for the conjugate momenta.

Errors

Return an error if the integration encounter a problem

Loading content...
Provided methods

fn integrate_symplectic(
&self,
l: &StateSync,
delta_t: Real
) -> Result<StateSync, Self::Error>

Integrate a Sync state by going to leap and then back to sync. This is the symplectic methode of integration, which should conserve the hamiltonian

Note that you might want to override this method as it can save you from a clone.

Errors

Return an error if the integration encounter a problem

Loading content...