Module lattice_qcd_rs::lattice

Defines lattices and lattice component.

LatticeCyclique is the structure that encode the lattice information like the lattice spacing, the number of point and the dimension. It is used to do operation on LatticePoint, LatticeLink and LatticeLinkCanonical. Or to get an iterator over these elements.

LatticePoint, LatticeLink and LatticeLinkCanonical are elements on the lattice. They encode where a field element is situated.

Structs

Direction
Represent a cardinal direction

IteratorDirection
Iterator over Direction with the same sign.

IteratorLatticeLinkCanonical
Iterator over LatticeLinkCanonical associated to a particular LatticeCyclique.

IteratorLatticePoint
Iterator over LatticePoint

LatticeCyclique
A cyclique lattice in space. Does not store point and links but is used to generate them.

LatticeLink
A lattice link, contrary to LatticeLinkCanonical the direction can be negative.

LatticeLinkCanonical
A canonical link of a lattice. It contain a position and a direction.

LatticePoint
Represents point on a (any) lattice.

Enums

DirectionEnum
Represent a cardinal direction

ErrorDirectionConversion
Error return by try from for Directions

IteratorElement
Enum for internal use of interator. It store the previous element returned by next

Traits

DirectionList
List all possible direction

LatticeElementToIndex
Trait to convert an element on a lattice to an usize.