Struct lattice_qcd_rs::lattice::Direction

pub struct Direction<const D: usize> { /* fields omitted */ }

Represent a cardinal direction

Implementations

impl<const D: usize> Direction<D>

pub const fn new(index_dir: usize, is_positive: bool) -> Option<Self>

New direction from a direction as an idex and wether it is in the positive direction.

pub fn positives_vec() -> Vec<Self>

List of all positives directions. This is very slow use Self::positive_directions instead

pub fn directions_vec() -> Vec<Self>

List all directions. This is very slow use DirectionList::get_all_directions instead

pub const fn get_directions<const IS_POSITIVE: bool>() -> [Self; D]

Get all direction with the sign IS_POSITIVE

pub const fn negative_directions() -> [Self; D]

Get all negative direction

pub const fn positive_directions() -> [Self; D]

Get all positive direction

pub const fn is_positive(&self) -> bool

Get if the position is positive.

pub const fn is_negative(&self) -> bool

Get if the position is Negative. see Direction::is_positive

pub const fn to_positive(self) -> Self

Return the positive direction associated, for example -x gives +x and +x gives +x.

Example

assert_eq!(
    Direction::<4>::new(1, false).unwrap().to_positive(),
    Direction::<4>::new(1, true).unwrap()
);
assert_eq!(
    Direction::<4>::new(1, true).unwrap().to_positive(),
    Direction::<4>::new(1, true).unwrap()
);

pub const fn to_index(&self) -> usize

Get a index associated to the direction.

Example

assert_eq!(Direction::<4>::new(1, false).unwrap().to_index(), 1);

pub fn to_vector(&self, a: f64) -> SVector<Real, D>

Convert the direction into a vector of norm a;

pub const fn dim() -> usize

Returns the dimension

pub fn to_unit_vector(&self) -> SVector<Real, D>

Convert the direction into a vector of norm 1;

pub fn from_vector(v: &SVector<Real, D>) -> Self

Find the direction the vector point the most. For a zero vector return DirectionEnum::XPos.

Example

assert_eq!(
    Direction::from_vector(&nalgebra::Vector4::new(1_f64, 0_f64, 0_f64, 0_f64)),
    Direction::<4>::new(0, true).unwrap()
);
assert_eq!(
    Direction::from_vector(&nalgebra::Vector4::new(0_f64, -1_f64, 0_f64, 0_f64)),
    Direction::<4>::new(1, false).unwrap()
);
assert_eq!(
    Direction::from_vector(&nalgebra::Vector4::new(0.5_f64, 1_f64, 0_f64, 2_f64)),
    Direction::<4>::new(3, true).unwrap()
);