Struct lattice_qcd_rs::lattice::LatticeCyclique

pub struct LatticeCyclique<const D: usize> { /* fields omitted */ }

A cyclique lattice in space. Does not store point and links but is used to generate them.

The generic parameter D is the dimension.

This lattice is cyclique more precisely if the lattice has N poits in each direction. Then we can mouve alongisde a direction going though point 0, 1, … N-1. The next step in the same direction goes back to the point at 0.

This structure is used for operation on LatticePoint, LatticeLink and LatticeLinkCanonical. access data on the lattice. These data are stored LinkMatrix and EField which are just a wrapper arround a Vec. LatticeCyclique is used to convert thes lattice element to an index to access these data.

This contain very few data and can be cloned at almost no cost even though it does not implement Copy.

Implementations

impl<const D: usize> LatticeCyclique<D>

pub const fn dim_st() -> usize

Number space + time dimension, this is the D parameter.

Not to confuse with LatticeCyclique::dim which is the number of point per dimension.

pub fn get_link_canonical(
&self,
pos: LatticePoint<D>,
dir: Direction<D>
) -> LatticeLinkCanonical<D>

see LatticeLinkCanonical, a conical link is a link whose direction is always positive. That means that a link form [x, y, z, t] with direction -x the link return is [x - 1, y, z, t] (modulo the lattice::dim()``) with direction +x`

Example

let lattice = LatticeCyclique::<4>::new(1_f64, 4).unwrap();
let point = LatticePoint::<4>::new([1, 0, 2, 0].into());
assert_eq!(
    lattice.get_link_canonical(point, DirectionEnum::XNeg.into()),
    LatticeLinkCanonical::new(LatticePoint::new([0, 0, 2, 0].into()), DirectionEnum::XPos.into()).unwrap()
);
assert_eq!(
    lattice.get_link_canonical(point, DirectionEnum::XPos.into()),
    LatticeLinkCanonical::new(LatticePoint::new([1, 0, 2, 0].into()), DirectionEnum::XPos.into()).unwrap()
);
assert_eq!(
    lattice.get_link_canonical(point, DirectionEnum::YNeg.into()),
    LatticeLinkCanonical::new(LatticePoint::new([1, 3, 2, 0].into()), DirectionEnum::YPos.into()).unwrap()
);

pub fn get_link(
&self,
pos: LatticePoint<D>,
dir: Direction<D>
) -> LatticeLink<D>

Return a link build from pos and dir.

It is similar to LatticeLink::new. It however enforce that the point is inside the bounds. If it is not, it will use the modulus of the bound.

Example

let l = LatticeCyclique::<3>::new(1_f64, 4).unwrap();
let dir = Direction::new(0, true).unwrap();
let pt = LatticePoint::new(SVector::<_, 3>::new(1, 2, 5));
let link = l.get_link(pt, dir);
assert_eq!(
    *link.pos(),
    LatticePoint::new(SVector::<_, 3>::new(1, 2, 1))
);

pub fn get_canonical(&self, l: LatticeLink<D>) -> LatticeLinkCanonical<D>

Transform a LatticeLink into a LatticeLinkCanonical.

See LatticeCyclique::get_link_canonical and LatticeLinkCanonical.

pub const fn dim(&self) -> usize

Get the number of points in a single direction.

use LatticeCyclique::get_number_of_points for the total number of points. Not to confuse with LatticeCyclique::dim_st which is the dimension of space-time.

pub fn get_points(&self) -> IteratorLatticePoint<'_, D>

Get an Iterator over all points of the lattice.

pub fn get_links(&self) -> IteratorLatticeLinkCanonical<'_, D>

Get an Iterator over all canonical link of the lattice.

pub fn new(size: Real, dim: usize) -> Result<Self, LatticeInitializationError>

create a new lattice with size the lattice size parameter, and dim the number of points in each spatial dimension.

Errors

Size should be greater than 0 and dim greater or equal to 2, otherwise return an error.

pub fn get_number_of_canonical_links_space(&self) -> usize

Total number of canonical links oriented in space for a set time.

basically the number of element return by LatticeCyclique::get_links

pub fn get_number_of_points(&self) -> usize

Total number of point in the lattice for a set time.

pub const fn size(&self) -> Real

Return the lattice size factor.

pub fn add_point_direction(
&self,
point: LatticePoint<D>,
dir: &Direction<D>
) -> LatticePoint<D>

get the next point in the lattice following the direction dir. It follows the cyclique property of the lattice.

Example

let lattice = LatticeCyclique::<4>::new(1_f64, 4).unwrap();
let point = LatticePoint::<4>::from([1, 0, 2, 0]);
assert_eq!(
    lattice.add_point_direction(point, &DirectionEnum::XPos.into()),
    LatticePoint::from([2, 0, 2, 0])
);
// In the following case we get [_, 3, _, _] because `dim = 4`, and this lattice is cyclique.
assert_eq!(
    lattice.add_point_direction(point, &DirectionEnum::YNeg.into()),
    LatticePoint::from([1, 3, 2, 0])
);

pub fn add_point_direction_n(
&self,
point: LatticePoint<D>,
dir: &Direction<D>,
shift_number: usize
) -> LatticePoint<D>

Returns the point given y mouving shift_number times in direction dir from position point. It follows the cyclique property of the lattice.

It is equivalent of doing Self::add_point_direction n times.

Example

let lattice = LatticeCyclique::<4>::new(1_f64, 4).unwrap();
let point = LatticePoint::<4>::from([1, 0, 2, 0]);
assert_eq!(
    lattice.add_point_direction_n(point, &DirectionEnum::XPos.into(), 2),
    LatticePoint::from([3, 0, 2, 0])
);
// In the following case we get [_, 1, _, _] because `dim = 4`, and this lattice is cyclique.
assert_eq!(
    lattice.add_point_direction_n(point, &DirectionEnum::YNeg.into(), 3),
    LatticePoint::from([1, 1, 2, 0])
);

pub fn has_compatible_lenght_links(&self, links: &LinkMatrix) -> bool

Retuns whether the number of canonical link is the same as the length of links

pub fn has_compatible_lenght_e_field(&self, e_field: &EField<D>) -> bool

Returns wether the number of point is the same as the length of e_field

pub fn has_compatible_lenght(
&self,
links: &LinkMatrix,
e_field: &EField<D>
) -> bool

Retuns the length is compatible for both links and e_field. See Self::has_compatible_lenght_links and Self::has_compatible_lenght_e_field.