Struct lattice_qcd_rs::lattice::LatticePoint

pub struct LatticePoint<const D: usize> { /* private fields */ }

Represents point on a (any) lattice.

Implementations

impl<const D: usize> LatticePoint<D>

pub const fn new(data: SVector<usize, D>) -> Self

Create a new lattice point.

It can be outside a lattice.

pub fn new_zero() -> Self

Create a point at the origin

pub fn from_fn<F>(f: F) -> Selfwhere F: FnMut(usize) -> usize,

Create a point using the closure generate elements with the index as input.

See nalgebra::base::Matrix::from_fn.

pub const fn len(&self) -> usize

Number of elements in LatticePoint. This is D.

pub const fn is_empty(&self) -> bool

Return if LatticePoint contain no data. True when the dimension is 0, false otherwise.

pub fn iter( &self ) -> impl Iterator<Item = &usize> + ExactSizeIterator + FusedIterator

Get an iterator on the data.

pub fn iter_mut( &mut self ) -> impl Iterator<Item = &mut usize> + ExactSizeIterator + FusedIterator

Get an iterator on the data as mutable.

pub const fn as_svector(&self) -> &SVector<usize, D>

Get the point as as nalgebra::SVector<usize, D>

Example

let max = point.as_svector().max();
let min = point.as_ref().min();

pub fn as_svector_mut(&mut self) -> &mut SVector<usize, D>

Get the point as a mut ref to nalgebra::SVector<usize, D>

Example

point.as_svector_mut()[2] = 2;
point.as_mut()[0] = 1;