Trait lattice_qcd_rs::lattice::LatticeElementToIndex

pub trait LatticeElementToIndex<const D: usize> {
    fn to_index(&self, l: &LatticeCyclique<D>) -> usize;
}

Trait to convert an element on a lattice to an usize.

Used mainly to index field on the lattice using std::vec::Vec

Required methods

fn to_index(&self, l: &LatticeCyclique<D>) -> usize

Given a lattice return an index from the element

Loading content...