Module lattice_qcd_rs::simulation::monte_carlo::hybride

Combine multiple Monte Carlo methods.

this module present different ways to combine multiple method

Structs

AdaptatorErrorMethod
Adaptator used to convert the error to another type. It is intented to use with HybrideMethodeVec.

HybrideMethodeCouple
This method can combine any two methods. The down side is that it can be very verbose to write Couples for a large number of methods.

HybrideMethodeVec
hybride methode that combine multiple methodes. It requires that all methods return the same error. You can use AdaptatorErrorMethod to convert the error. If you want type with different error you can use HybrideMethodeCouple.

Enums

HybrideMethodeCoupleError
Error given by HybrideMethodeCouple

HybrideMethodeVecError
Error given by HybrideMethodeVec

Type Definitions

HybrideMethodeQuadruple
Combine four methods.

HybrideMethodeQuadrupleError
Error returned by HybrideMethodeQuadruple.

HybrideMethodeQuintuple
Combine four methods.

HybrideMethodeTriple
Combine three methods.

HybrideMethodeTripleError
Error returned by HybrideMethodeTriple.

HybrideMethodeTripleQuintuple
Error returned by HybrideMethodeQuintuple.