Struct lattice_qcd_rs::simulation::monte_carlo::metropolis_hastings::MetropolisHastingsDiagnostic

pub struct MetropolisHastingsDiagnostic { /* private fields */ }

Metropolis Hastings algorithm with diagnostics. Very slow, use MetropolisHastingsDeltaDiagnostic instead.

Similar to MetropolisHastingsDiagnostic but with diagnostic information.

Note that this method does not do a sweep but change random link matrix, for a sweep there is super::MetropolisHastingsSweep.

Example

see example of super::McWrapper

Implementations

impl MetropolisHastingsDiagnostic

pub fn new(number_of_update: usize, spread: Real) -> Option<Self>

spread should be between 0 and 1 both not included and number_of_update should be greater than 0. 0.1_f64 is a good choice for this parameter.

number_of_update is the number of times a link matrix is randomly changed. spread is the spread factor for the random matrix change ( used in su3::random_su3_close_to_unity).

pub const fn prob_replace_last(&self) -> Real

Get the last probably of acceptance of the random change.

pub const fn has_replace_last(&self) -> bool

Get if last step has accepted the replacement.

pub const fn number_of_update(&self) -> usize

Get the number of updates per steps.

pub const fn spread(&self) -> Real

Get the spread parameter.