Struct lattice_qcd_rs::simulation::monte_carlo::metropolis_hastings_sweep::MetropolisHastingsSweep

pub struct MetropolisHastingsSweep<Rng: Rng> { /* private fields */ }

Metropolis Hastings method by doing a pass on all points

Example

use lattice_qcd_rs::error::ImplementationError;
use lattice_qcd_rs::simulation::{LatticeState, LatticeStateDefault, MetropolisHastingsSweep};
use rand::SeedableRng;

let rng = rand::rngs::StdRng::seed_from_u64(0); // change with your seed
let mut mh = MetropolisHastingsSweep::new(1, 0.1_f64, rng)
    .ok_or(ImplementationError::OptionWithUnexpectedNone)?;
// Realistically you want more steps than 10

let mut state = LatticeStateDefault::<3>::new_cold(1_f64, 6_f64, 4)?;
for _ in 0..10 {
    state = state.monte_carlo_step(&mut mh)?;
    println!(
        "mean probability of acceptance during last step {} and replaced {} links",
        mh.prob_replace_mean(),
        mh.number_replace_last()
    );
    // operation to track the progress or the evolution
}
// operation at the end of the simulation

Implementations

impl<Rng: Rng> MetropolisHastingsSweep<Rng>

pub const fn rng(&self) -> &Rng

Get a ref to the rng.

pub fn new(number_of_update: usize, spread: Real, rng: Rng) -> Option<Self>

spread should be between 0 and 1 both not included and number_of_update should be greater than 0.

number_of_update is the number of times a link matrix is randomly changed. spread is the spread factor for the random matrix change ( used in su3::random_su3_close_to_unity).

pub const fn prob_replace_mean(&self) -> Real

Get the mean of last probably of acceptance of the random change.

pub const fn number_replace_last(&self) -> usize

Number of accepted change during last sweep

pub fn rng_owned(self) -> Rng

Get the last probably of acceptance of the random change.

pub fn rng_mut(&mut self) -> &mut Rng

Get a mutable reference to the rng.