Trait lattice_qcd_rs::simulation::monte_carlo::MonteCarloDefault

pub trait MonteCarloDefault<State, const D: usize> where
    State: LatticeState<D>, {
    type Error;
    fn get_potential_next_element(
        &mut self, 
        state: &State, 
        rng: &mut impl Rng
    ) -> Result<State, Self::Error>;

    fn get_probability_of_replacement(
        old_state: &State, 
        new_state: &State
    ) -> Real { ... }

    fn get_next_element_default(
        &mut self, 
        state: State, 
        rng: &mut impl Rng
    ) -> Result<State, Self::Error> { ... }
}

Some times is is esayer to just implement a potential next element, the rest is done automatically.

To get an MonteCarlo use the wrapper McWrapper

Associated Types

type Error

Error returned while getting the next ellement.

Loading content...
Required methods

fn get_potential_next_element(
&mut self,
state: &State,
rng: &mut impl Rng
) -> Result<State, Self::Error>

Generate a radom element from the previous element ( like a Markov chain).

Errors

Gives an error if a potential next ellement cannot be generated.

Loading content...
Provided methods

fn get_probability_of_replacement(old_state: &State, new_state: &State) -> Real

probability of the next element to replace the current one.

by default it is Exp(-H_old) / Exp(-H_new).

fn get_next_element_default(
&mut self,
state: State,
rng: &mut impl Rng
) -> Result<State, Self::Error>

Get the next element in the chain either the old state or a new one replacing it.

Errors

Gives an error if a potential next ellement cannot be generated.

Loading content...