Module lattice_qcd_rs::simulation::state

Module containing the simulation State

Structs

LatticeStateDefault
Represent a simulation state at a set time.

LatticeStateWithEFieldSyncDefault
wrapper to implement LatticeStateWithEField from a LatticeState using the default implementation of conjugate momenta.

SimulationStateLeap
wrapper for a simulation state using leap frog (SimulationStateLeap) using a synchrone type (SimulationStateSynchrone).

Traits

LatticeState
trait to represent a pure gauge lattice state.

LatticeStateNew
trait for a way to create a LatticeState from some parameters.

LatticeStateWithEField
Represent a lattice state where the conjugate momenta of the link matrices are included.

LatticeStateWithEFieldNew
Trait to create a simulation state

SimulationStateLeapFrog
LatticeStateWithEField who represent link matrices at time T and its conjugate momenta at time T + 1/2.

SimulationStateSynchrone
LatticeStateWithEField who represent link matrices at the same time position as its conjugate momenta e_field.

Type Definitions

LeapFrogStateDefault
Default leap frog simulation state