Struct lattice_qcd_rs::simulation::state::LatticeStateDefault

pub struct LatticeStateDefault<const D: usize> { /* fields omitted */ }

Represent a simulation state at a set time.

It has the default pure gauge hamiltonian

Implementations

impl<const D: usize> LatticeStateDefault<D>

pub fn new_cold(
size: Real,
beta: Real,
number_of_points: usize
) -> Result<Self, StateInitializationError>

Create a cold configuration. i.e. all the links are set to the unit matrix.

With the lattice of size size and dimension number_of_points ( see LatticeCyclique::new ) and beta parameter beta.

Errors

Returns StateInitializationError::LatticeInitializationError if the parameter is invalide for LatticeCyclique. Or propagate the error form Self::new.

pub fn new_deterministe(
size: Real,
beta: Real,
number_of_points: usize,
rng: &mut impl Rng
) -> Result<Self, StateInitializationError>

Create a “hot” configuration, i.e. the link matrices are chosen randomly.

With the lattice of size size and dimension number_of_points ( see LatticeCyclique::new ) and beta parameter beta.

The creation is determiste in the sence that it is reproducible:

Errors

Returns StateInitializationError::LatticeInitializationError if the parameter is invalide for LatticeCyclique. Or propagate the error form Self::new.

Example

This example demonstrate how to reproduce the same configuration

extern crate rand;
extern crate rand_distr;
use rand::{rngs::StdRng, SeedableRng};

let mut rng_1 = StdRng::seed_from_u64(0);
let mut rng_2 = StdRng::seed_from_u64(0);
// They have the same seed and should generate the same numbers
assert_eq!(
    LatticeStateDefault::<4>::new_deterministe(1_f64, 1_f64, 4, &mut rng_1).unwrap(),
    LatticeStateDefault::<4>::new_deterministe(1_f64, 1_f64, 4, &mut rng_2).unwrap()
);

pub fn normalize_link_matrices(&mut self)

Correct the numerical drift, reprojecting all the link matrices to SU(3). see LinkMatrix::normalize.

pub fn get_link_mut(
&mut self,
link: &LatticeLinkCanonical<D>
) -> Option<&mut CMatrix3>

Get a mutable reference to the link matrix at link

pub fn link_matrix_owned(self) -> LinkMatrix

Absorbe self anf return the link_matrix as owned