Struct lattice_qcd_rs::simulation::state::LatticeStateEFSyncDefault

pub struct LatticeStateEFSyncDefault<State, const D: usize>where
    State: LatticeState<D> + ?Sized,{ /* private fields */ }

wrapper to implement LatticeStateWithEField from a LatticeState using the default implementation of conjugate momenta.

It also implement SimulationStateSynchronous.

Implementations

impl<State, const D: usize> LatticeStateEFSyncDefault<State, D>where State: LatticeState<D> + ?Sized,

pub fn state_owned(self) -> Statewhere State: Sized,

Absorbs self and return the state as owned. It essentially deconstruct the structure.

pub const fn lattice_state(&self) -> &State

Get a reference to the state.

pub fn lattice_state_mut(&mut self) -> &mut State

Get a mutable reference to the state.

pub fn new_random_e_state(lattice_state: State, rng: &mut impl Rng) -> Selfwhere State: Sized,

Take a state and generate a new random one and try projecting it to the Gauss law.

Panic

Panics if N(0, 0.5/beta ) is not a valid distribution (for example beta = 0). Panics if the field could not be projected to the Gauss law.

pub fn new_e_cold(lattice_state: State) -> Selfwhere State: Sized,

Create a new Self from a state and a cold configuration of the e field (i.e. set to 0)

pub fn e_field_mut(&mut self) -> &mut EField<D>

Get a mutable reference to the efield

impl<State, const D: usize> LatticeStateEFSyncDefault<State, D>where Self: LatticeStateWithEField<D>, State: LatticeState<D> + ?Sized,

pub fn gauss(&self, point: &LatticePoint<D>) -> Option<CMatrix3>

Get the gauss coefficient G(x) = \sum_i E_i(x) - U_{-i}(x) E_i(x - i) U^\dagger_{-i}(x).

impl<State, const D: usize> LatticeStateEFSyncDefault<State, D>where Self: LatticeStateWithEFieldNew<D>, <Self as LatticeStateWithEFieldNew<D>>::Error: From<LatticeInitializationError>, State: LatticeState<D>,

pub fn new_determinist<R>( size: Real, beta: Real, number_of_points: usize, rng: &mut R, d: &impl Distribution<Real> ) -> Result<Self, <Self as LatticeStateWithEFieldNew<D>>::Error>where R: Rng + ?Sized,

Generate a hot (i.e. random) initial state.

Single threaded generation with a given random number generator. size is the size parameter of the lattice and number_of_points is the number of points in each spatial dimension of the lattice. See LatticeCyclic::new for more info.

useful to reproduce a set of data but slower than LatticeStateEFSyncDefault::new_random_threaded.

Errors

Return StateInitializationError::LatticeInitializationError if the parameter is invalid for LatticeCyclic. Or propagates the error form Self::new.

Example

use rand::{SeedableRng,rngs::StdRng};

let mut rng_1 = StdRng::seed_from_u64(0);
let mut rng_2 = StdRng::seed_from_u64(0);
// They have the same seed and should generate the same numbers
let distribution = rand::distributions::Uniform::from(-1_f64..1_f64);
assert_eq!(
    LatticeStateEFSyncDefault::<LatticeStateDefault<4>, 4>::new_determinist(1_f64, 1_f64, 4, &mut rng_1, &distribution).unwrap(),
    LatticeStateEFSyncDefault::<LatticeStateDefault<4>, 4>::new_determinist(1_f64, 1_f64, 4, &mut rng_2, &distribution).unwrap()
);

pub fn new_determinist_cold_e_hot_link<R>( size: Real, beta: Real, number_of_points: usize, rng: &mut R ) -> Result<Self, <Self as LatticeStateWithEFieldNew<D>>::Error>where R: Rng + ?Sized,

Generate a configuration with cold e_field and hot link matrices

Errors

Return StateInitializationError::LatticeInitializationError if the parameter is invalid for LatticeCyclic. Or propagates the error form Self::new.

pub fn new_cold( size: Real, beta: Real, number_of_points: usize ) -> Result<Self, <Self as LatticeStateWithEFieldNew<D>>::Error>

Generate a new cold state.

It meas that the link matrices are set to the identity and electrical field are set to 0.

Errors

Return StateInitializationError::LatticeInitializationError if the parameter is invalid for LatticeCyclic. Or propagates the error form Self::new.

impl<State, const D: usize> LatticeStateEFSyncDefault<State, D>where Self: LatticeStateWithEFieldNew<D, Error = StateInitializationError>, State: LatticeState<D>,

pub fn new_random_threaded<Distribution>( size: Real, beta: Real, number_of_points: usize, d: &Distribution, number_of_thread: usize ) -> Result<Self, ThreadedStateInitializationError>where Distribution: Distribution<Real> + Sync,

Generate a hot (i.e. random) initial state.

Multi threaded generation of random data. Due to the non deterministic way threads operate a set cannot be reproduce easily, In that case use LatticeStateEFSyncDefault::new_determinist.

Errors

Return StateInitializationError::LatticeInitializationError if the parameter is invalid for LatticeCyclic. Return ThreadError::ThreadNumberIncorrect if number_of_points = 0. Returns an error if a thread panicked. Finally, propagates the error form Self::new.