Struct lattice_qcd_rs::simulation::state::SimulationStateLeap

pub struct SimulationStateLeap<State, const D: usize> where
    State: SimulationStateSynchrone<D>,  { /* fields omitted */ }

wrapper for a simulation state using leap frog (SimulationStateLeap) using a synchrone type (SimulationStateSynchrone).

Implementations

impl<State, const D: usize> SimulationStateLeap<State, D> where
State: SimulationStateSynchrone<D> + LatticeStateWithEField<D>,

pub fn state(&self) -> &State

get a reference to the state

pub fn state_mut(&mut self) -> &mut State

get a mutable reference to the state

pub fn from_synchrone<I: ?Sized>(
s: &State,
integrator: &I,
delta_t: Real
) -> Result<Self, I::Error> where
I: SymplecticIntegrator<State, Self, D>,

Create a leap state from a sync one by integrating by half a setp the e_field.

Errors

Returns an error if the intregration failed.

pub fn get_gauss(&self, point: &LatticePoint<D>) -> Option<CMatrix3>

Get the gauss coefficient G(x) = \sum_i E_i(x) - U_{-i}(x) E_i(x - i) U^\dagger_{-i}(x).