Trait lattice_qcd_rs::simulation::state::LatticeState

pub trait LatticeState<const D: usize> where
    Self: Sync + Sized + Debug, {
    const CA: Real;

    fn link_matrix(&self) -> &LinkMatrix;

    fn set_link_matrix(&mut self, link_matrix: LinkMatrix);

    fn lattice(&self) -> &LatticeCyclique<D>;

    fn beta(&self) -> Real;

    fn get_hamiltonian_links(&self) -> Real;

    fn monte_carlo_step<M: ?Sized>(self, m: &mut M) -> Result<Self, M::Error>
    where
        M: MonteCarlo<Self, D>,
    { ... }

    fn average_trace_plaquette(&self) -> Option<Complex> { ... }
}

trait to represent a pure gauge lattice state.

It defines only one field: link_matrix of type LinkMatrix.

Associated Constants

const CA: Real

C_A constant of the model

Loading content...
Required methods

fn link_matrix(&self) -> &LinkMatrix

The link matrices of this state.

fn set_link_matrix(&mut self, link_matrix: LinkMatrix)

Replace the links matrices with the given input. It should panic if link matrix is not of the correct size.

Panic

Panic if the length of link_matrix is different from lattice.get_number_of_canonical_links_space()

fn lattice(&self) -> &LatticeCyclique<D>

get the lattice into which the state exists

fn beta(&self) -> Real

return the beta parameter of the states

fn get_hamiltonian_links(&self) -> Real

return the Hamiltonian of the links configuration

Loading content...
Provided methods

fn monte_carlo_step<M: ?Sized>(self, m: &mut M) -> Result<Self, M::Error> where
M: MonteCarlo<Self, D>,

Do one monte carlo step with the given

Errors

Get an error if the simulation could not be advantaced

fn average_trace_plaquette(&self) -> Option<Complex>

Take the average of the trace of all plaquettes

Loading content...