Trait lattice_qcd_rs::simulation::state::LatticeStateWithEField

pub trait LatticeStateWithEField<const D: usize> where
    Self: Sized + Sync + LatticeState<D> + Debug, {
    fn e_field(&self) -> &EField<D>;

    fn set_e_field(&mut self, e_field: EField<D>);

    fn t(&self) -> usize;

    fn get_derivative_u(
        link: &LatticeLinkCanonical<D>, 
        link_matrix: &LinkMatrix, 
        e_field: &EField<D>, 
        lattice: &LatticeCyclique<D>
    ) -> Option<CMatrix3>;

    fn get_derivative_e(
        point: &LatticePoint<D>, 
        link_matrix: &LinkMatrix, 
        e_field: &EField<D>, 
        lattice: &LatticeCyclique<D>
    ) -> Option<SVector<Su3Adjoint, D>>;

    fn get_hamiltonian_efield(&self) -> Real;

    fn reset_e_field(
        &mut self, 
        rng: &mut impl Rng
    ) -> Result<(), StateInitializationError> { ... }

    fn get_hamiltonian_total(&self) -> Real { ... }
}

Represent a lattice state where the conjugate momenta of the link matrices are included.

If you have a LatticeState and want the default way of adding the conjugate momenta look at LatticeStateWithEFieldSyncDefault.

If you want to solve the equation of motion using an SymplecticIntegrator also implement SimulationStateSynchrone and SimulationStateLeap can give you an SimulationStateLeapFrog.

It is used for the super::monte_carlo::HybridMonteCarlo algorithm.

Required methods

fn e_field(&self) -> &EField<D>

The “Electrical” field of this state.

fn set_e_field(&mut self, e_field: EField<D>)

Replace the electrical field with the given input. It should panic if the input is not of the correct size.

Panic

Panic if the length of link_matrix is different from lattice.get_number_of_points()

fn t(&self) -> usize

return the time state, i.e. the number of time the simulation ran.

fn get_derivative_u(
link: &LatticeLinkCanonical<D>,
link_matrix: &LinkMatrix,
e_field: &EField<D>,
lattice: &LatticeCyclique<D>
) -> Option<CMatrix3>

get the derivative \partial_t U(link)

fn get_derivative_e(
point: &LatticePoint<D>,
link_matrix: &LinkMatrix,
e_field: &EField<D>,
lattice: &LatticeCyclique<D>
) -> Option<SVector<Su3Adjoint, D>>

get the derivative \partial_t E(point)

fn get_hamiltonian_efield(&self) -> Real

Get the energy of the conjugate momenta configuration

Loading content...
Provided methods

fn reset_e_field(
&mut self,
rng: &mut impl Rng
) -> Result<(), StateInitializationError>

Reset the e_field with radom value distributed as N(0, 1/beta ) rand_distr::StandardNormal.

Errors

Gives and error if N(0, 0.5/beta ) is not a valide distribution (for exampple beta = 0)

fn get_hamiltonian_total(&self) -> Real

Get the total energy, by default LatticeStateWithEField::get_hamiltonian_efield + LatticeState::get_hamiltonian_links

Loading content...