Trait lattice_qcd_rs::simulation::state::SimulationStateLeapFrog

pub trait SimulationStateLeapFrog<const D: usize> where
    Self: LatticeStateWithEField<D>, {
    fn simulate_to_synchrone<I: ?Sized, State>(
        &self, 
        integrator: &I, 
        delta_t: Real
    ) -> Result<State, I::Error>
    where
        State: SimulationStateSynchrone<D>,
        I: SymplecticIntegrator<State, Self, D>,
    { ... }

    fn simulate_leap<I: ?Sized, T>(
        &self, 
        integrator: &I, 
        delta_t: Real
    ) -> Result<Self, I::Error>
    where
        I: SymplecticIntegrator<T, Self, D>,
        T: SimulationStateSynchrone<D>,
    { ... }

    fn simulate_leap_n<I: ?Sized, T>(
        &self, 
        integrator: &I, 
        delta_t: Real, 
        numbers_of_times: usize
    ) -> Result<Self, MultiIntegrationError<I::Error>>
    where
        I: SymplecticIntegrator<T, Self, D>,
        T: SimulationStateSynchrone<D>,
    { ... }
}

LatticeStateWithEField who represent link matrices at time T and its conjugate momenta at time T + 1/2.

If you have a SimulationStateSynchrone look at the wrapper SimulationStateLeap.

Provided methods

fn simulate_to_synchrone<I: ?Sized, State>(
&self,
integrator: &I,
delta_t: Real
) -> Result<State, I::Error> where
State: SimulationStateSynchrone<D>,
I: SymplecticIntegrator<State, Self, D>,

Simulate the state to synchrone by finishing the half setp.

Errors

Return an error if the integration could not be done.

fn simulate_leap<I: ?Sized, T>(
&self,
integrator: &I,
delta_t: Real
) -> Result<Self, I::Error> where
I: SymplecticIntegrator<T, Self, D>,
T: SimulationStateSynchrone<D>,

Does one simulation step using the leap frog algorithm.

Errors

Return an error if the integration could not be done.

fn simulate_leap_n<I: ?Sized, T>(
&self,
integrator: &I,
delta_t: Real,
numbers_of_times: usize
) -> Result<Self, MultiIntegrationError<I::Error>> where
I: SymplecticIntegrator<T, Self, D>,
T: SimulationStateSynchrone<D>,

does numbers_of_times simulation set of size delta_t using the leap frog algorithm.

Errors

Return an error if the integration could not be done or MultiIntegrationError::ZeroIntegration is the number of step is zero.

Loading content...