Module lattice_qcd_rs::statistics

provide statistical tools

Re-exports

pub use distribution::*;

Modules

distribution
Some statistical distribution used by other part of the library

Functions

covariance
compute the covariance between two slices. Return None if the slices are not of the same length

covariance_par_iter
compute the covariance between two rayon::iter::IndexedParallelIterator. Return None if the par iters are not of the same length

covariance_par_iter_val
compute the covariance between two rayon::iter::IndexedParallelIterator by value, Return None if the par iters are not of the same length

mean
compute the mean from a slice

mean_and_variance
compute the mean and variance (squared of standard deviation) from a slice

mean_and_variance_par_iter
Compute the mean and variance (squared of standard deviation) from a rayon::iter::IndexedParallelIterator.

mean_and_variance_par_iter_val
Compute the mean and variance (squared of standard deviation) from a rayon::iter::IndexedParallelIterator by value.

mean_par_iter
Compute the mean from a rayon::iter::IndexedParallelIterator.

mean_par_iter_val
Compute the mean from a rayon::iter::IndexedParallelIterator by value.

mean_with_error
compute the mean the statistocal error on this value a slice

mean_with_error_par_iter
compute the mean the statistocal error on this value a rayon::iter::IndexedParallelIterator.

mean_with_error_par_iter_val
compute the mean the statistocal error on this value a rayon::iter::IndexedParallelIterator by value.

variance
compute the variance (squared of standard deviation) from a slice

variance_par_iter
Compute the variance (squared of standard deviation) from a rayon::iter::IndexedParallelIterator.

variance_par_iter_val
Compute the variance (squared of standard deviation) from a rayon::iter::IndexedParallelIterator by value.