Module lattice_qcd_rs::su2

module for SU(2) matrix

Constants

PAULI_1
First Pauli matrix.

PAULI_2
Second Pauli matrix.

PAULI_3
Third Pauli matrix.

PAULI_MATRICES
List of Pauli matrices, see wikipedia

Functions

get_complex_matrix_from_vec
Return x0 1 + i x_i * \sigma_i.

get_random_su2
Get an Uniformly random SU(2) matrix.

get_random_su2_close_to_unity
Get a radom SU(2) matrix close the 1 or -1.

is_matrix_su2
Return wether the input matrix is SU(2) up to epsilon.

project_to_su2
Project the matrix to SU(2). Return the identity if the norm after unormalize is subnormal (seef64::is_normal).

project_to_su2_unorm
Take any 2x2 matrix and project it to a matric X such that X / X.determinant().modulus().sqrt() is SU(2).