Module lattice_qcd_rs::su3

Module for SU(3) matrices and su(3) (that is the generators of SU(3) )

The module defines the SU(3) generator we use the same matrices as on wikipedia divided by two such that Tr(T^a T^b) = \delta^{ab} /2 .

Constants

GENERATORS
list of SU(3) generators they are normalize such that Tr(T^a T^b) = \frac{1}{2}\delta^{ab}

GENERATOR_1
SU(3) generator

GENERATOR_2
SU(3) generator

GENERATOR_3
SU(3) generator

GENERATOR_4
SU(3) generator

GENERATOR_5
SU(3) generator

GENERATOR_6
SU(3) generator

GENERATOR_7
SU(3) generator

GENERATOR_8
SU(3) generator

Traits

MatrixExp
Exponential of matrices.

Functions

extract_su2_unorm
Get the three unormalize sub SU(2) matrix of the given SU(3) matrix, ordered r, s, t

get_factorial_size_for_exp
Return N such that 1/(N-7)! < f64::EPSILON.

get_r
Embed a Matrix2 inside Matrix3 leaving the last row and column be the same as identity.

get_random_su3
Generate Uniformly distributed SU(3)

get_random_su3_close_to_unity
Get a radom SU(3) matrix close to [get_r] (+/- 1) * [get_s] (+/- 1) * [get_t] (+/- 1).

get_s
Embed a Matrix2 inside Matrix3 leaving the second row and column be the same as identity.

get_su2_r_unorm
Get the unormalize SU(2) sub matrix of an SU(3) matrix correspondig to the “r” sub block see get_sub_block_r and get_r.

get_su2_s_unorm
Get the unormalize SU(2) sub matrix of an SU(3) matrix correspondig to the “s” sub block see get_sub_block_s and get_s.

get_su2_t_unorm
Get the unormalize SU(2) sub matrix of an SU(3) matrix correspondig to the “t” sub block see get_sub_block_t and get_t.

get_sub_block_r
get the CMatrix2 sub block corresponing to get_r

get_sub_block_s
get the CMatrix2 sub block corresponing to get_s

get_sub_block_t
get the CMatrix2 sub block corresponing to get_t

get_t
Embed a Matrix2 inside Matrix3 leaving the first row and column be the same as identity.

is_matrix_su3
return wether the input matrix is SU(3) up to epsilon.

is_matrix_su3_lie
Returns wether the given matric is in the lie algebra su(3) that generates SU(3) up to epsilon

matrix_su3_exp_i
the input must be a su(3) matrix (generator of SU(3)), gives the SU3 matrix from the adjoint rep, i.e compute exp(i v^a T^a )

matrix_su3_exp_r
the input must be a su(3) matrix (generator of SU(3)), gives the value exp(v^a T^a )

orthonormalize_matrix
Try orthonormalize the given matrix.

orthonormalize_matrix_mut
Orthonormalize the given matrix by mutating its content.

reverse
Return the matrix

su3_exp_i
give the SU3 matrix from the adjoint rep, i.e compute exp(i v^a T^a )

su3_exp_r
gives the value exp(v^a T^a )