Function lattice_qcd_rs::thread::run_pool_parallel

pub fn run_pool_parallel<Key, Data, CommonData, F>(
    iter: impl Iterator<Item = Key> + Send, 
    common_data: &CommonData, 
    closure: &F, 
    number_of_thread: usize, 
    capacity: usize
) -> Result<HashMap<Key, Data>, ThreadError> where
    CommonData: Sync,
    Key: Eq + Hash + Send + Clone + Sync,
    Data: Send,
    F: Sync + Clone + Fn(&Key, &CommonData) -> Data, 

run jobs in parallel.

The pool of job is given by iter. the job is given by closure that have the form |key,common_data| -> Data. number_of_thread determine the number of job done in parallel and should be greater than 0, otherwise return ThreadError::ThreadNumberIncorect. capacity is used to determine the capacity of the HashMap upon initisation (see HashMap::with_capacity)

Errors

Returns ThreadError::ThreadNumberIncorect is the number of threads is 0. Returns ThreadError::Panic if a thread panicked. Containt the panick message.

Example

let us computes the value of i^2 * c for i in [2,9999] with 4 threads

let iter = 2..10000;
let c = 5;
// we could have put 4 inside the closure but this demonstrate how to use common data
let result = run_pool_parallel(iter, &c, &|i, c| i * i * c, 4, 10000 - 2).unwrap();
assert_eq!(*result.get(&40).unwrap(), 40 * 40 * c);
assert_eq!(result.get(&1), None);

In the next example a thread will panic, we demonstrate the return type.

let iter = 0..10;
let result = run_pool_parallel(iter, &(), &|_, _| panic!("panic message"), 4, 10);
result.unwrap(); // this propagate the panic.

This give the following panic message

---- src\thread.rs - thread::run_pool_parallel (line 51) stdout ----
Test executable failed (exit code 101).

stderr:
thread '<unnamed>' panicked at 'thread 'panic message', src\thread.rs:6:60
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
<unnamed>' panicked at 'panic message', src\thread.rs:6:60
thread '<unnamed>' panicked at 'panic message', src\thread.rs:6:60
thread '<unnamed>' panicked at 'panic message', src\thread.rs:6:60
thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: Panic(Any)', src\thread.rs:7:8