Module lattice_qcd_rs::error

Defines different error types.

Structs

ErrorWithOwnedValue

A struct that combine an error with a owned value

Enums

ImplementationError

Errors in the implementation of the library. This is unwanted to return this type but sometimes this is better to return that instead of panicking. It is also used in some example.

LatticeInitializationError

Error while initializing a lattice

MultiIntegrationError

Error return while doing multiple steps.

Never

A type that can never be (safely) initialized. This is temporary, until never is accepted into stable rust.

StateInitializationError

Error while initializing a state

ThreadedStateInitializationError

Error while initializing a state using multiple thread or threaded function.