Enum lattice_qcd_rs::lattice::DirectionEnum

pub enum DirectionEnum {
    XPos,
    XNeg,
    YPos,
    YNeg,
    ZPos,
    ZNeg,
    TPos,
    TNeg,
}

Represent a cardinal direction

Variants

XPos

Positive x direction.

XNeg

Negative x direction.

YPos

Positive y direction.

YNeg

Negative y direction.

ZPos

Positive z direction.

ZNeg

Negative z direction.

TPos

Positive t direction.

TNeg

Negative t direction.

Implementations

impl DirectionEnum

pub const DIRECTIONS: [Self; 8] = _

List all directions.

pub const DIRECTIONS_SPACE: [Self; 6] = _

List all spatial directions.

pub const POSITIVES: [Self; 4] = _

List of all positives directions.

pub const POSITIVES_SPACE: [Self; 3] = _

List spatial positive direction.

pub fn to_vector(self, a: f64) -> Vector4<Real>

Convert the direction into a vector of norm a;

pub const fn to_unit_vector(self) -> Vector4<Real>

Convert the direction into a vector of norm 1;

pub const fn is_positive(self) -> bool

Get if the position is positive.

Example

assert_eq!(DirectionEnum::XPos.is_positive(), true);
assert_eq!(DirectionEnum::TPos.is_positive(), true);
assert_eq!(DirectionEnum::YNeg.is_positive(), false);

pub const fn is_negative(self) -> bool

Get if the position is Negative. see DirectionEnum::is_positive

pub fn from_vector(v: &Vector4<Real>) -> Self

Find the direction the vector point the most. For a zero vector return DirectionEnum::XPos.

Example

assert_eq!(
    DirectionEnum::from_vector(&nalgebra::Vector4::new(1_f64, 0_f64, 0_f64, 0_f64)),
    DirectionEnum::XPos
);
assert_eq!(
    DirectionEnum::from_vector(&nalgebra::Vector4::new(0_f64, -1_f64, 0_f64, 0_f64)),
    DirectionEnum::YNeg
);
assert_eq!(
    DirectionEnum::from_vector(&nalgebra::Vector4::new(0.5_f64, 1_f64, 0_f64, 2_f64)),
    DirectionEnum::TPos
);

pub const fn to_positive(self) -> Self

Return the positive direction associated, for example -x gives +x and +x gives +x.

Example

assert_eq!(DirectionEnum::XNeg.to_positive(), DirectionEnum::XPos);
assert_eq!(DirectionEnum::YPos.to_positive(), DirectionEnum::YPos);

pub const fn to_index(self) -> usize

Get a index associated to the direction.

Example

assert_eq!(DirectionEnum::XPos.to_index(), 0);
assert_eq!(DirectionEnum::XNeg.to_index(), 0);
assert_eq!(DirectionEnum::YPos.to_index(), 1);
assert_eq!(DirectionEnum::YNeg.to_index(), 1);
assert_eq!(DirectionEnum::ZPos.to_index(), 2);
assert_eq!(DirectionEnum::ZNeg.to_index(), 2);
assert_eq!(DirectionEnum::TPos.to_index(), 3);
assert_eq!(DirectionEnum::TNeg.to_index(), 3);