Module lattice_qcd_rs::simulation::state

Module containing the simulation State.

The trait LatticeState is the most basic state with just the link matrices. The default implementation is LatticeStateDefault. More over a more advance state with the color electrical field is given by the trait LatticeStateWithEField with the default implementation LatticeStateEFSyncDefault. Thw word “Sync” appears. It means that the simulation for the equation of movement are done at the same time for both the link matrices and the electric field. It is given by te trait SimulationStateSynchronous. It can also be done using the leapfrog method with the trait SimulationStateLeapFrog and the wrapper SimulationStateLeap.

See item documentation for more details.

Structs

LatticeStateDefault

Represent a simulation state at a set time.

LatticeStateEFSyncDefault

wrapper to implement LatticeStateWithEField from a LatticeState using the default implementation of conjugate momenta.

SimulationStateLeap

wrapper for a simulation state using leap frog (SimulationStateLeap) using a synchronous type (SimulationStateSynchronous).

Traits

LatticeState

Trait to represent a pure gauge lattice state of dimension D.

LatticeStateNew

Trait for a way to create a LatticeState from some parameters.

LatticeStateWithEField

Represent a lattice state where the conjugate momenta of the link matrices are included.

LatticeStateWithEFieldNew

Trait to create a simulation state.

SimulationStateLeapFrog

LatticeStateWithEField who represent link matrices at time T and its conjugate momenta at time T + 1/2.

SimulationStateSynchronous

LatticeStateWithEField who represent link matrices at the same time position as its conjugate momenta e_field.

Type Definitions

LeapFrogStateDefault

Default leap frog simulation state