Struct lattice_qcd_rs::simulation::state::SimulationStateLeap

pub struct SimulationStateLeap<State, const D: usize>where
    State: SimulationStateSynchronous<D> + ?Sized,{ /* private fields */ }

wrapper for a simulation state using leap frog (SimulationStateLeap) using a synchronous type (SimulationStateSynchronous).

Implementations

impl<State, const D: usize> SimulationStateLeap<State, D>where State: SimulationStateSynchronous<D> + LatticeStateWithEField<D> + ?Sized,

pub const fn state(&self) -> &State

get a reference to the state

pub const fn new_from_state(state: State) -> Self

Create a new SimulationStateLeap directly from a state without applying any modification.

In most cases wou will prefer to build it using LatticeStateNew or Self::from_synchronous.

pub fn state_mut(&mut self) -> &mut State

get a mutable reference to the state

pub fn from_synchronous<I>( s: &State, integrator: &I, delta_t: Real ) -> Result<Self, I::Error>where I: SymplecticIntegrator<State, Self, D> + ?Sized,

Create a leap state from a sync one by integrating by half a step the e_field.

Errors

Returns an error if the integration failed.

pub fn gauss(&self, point: &LatticePoint<D>) -> Option<CMatrix3>

Get the gauss coefficient G(x) = \sum_i E_i(x) - U_{-i}(x) E_i(x - i) U^\dagger_{-i}(x).